pyproject.toml
[tool.poetry]
name = "recon"
version = "0.1.0"
description = ""
authors = ["adriyansyahmf <adriyansyahmf0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^21.4.0"
fastapi = "^0.78.0"
api = "^0.0.7"
python-whois = "^0.7.3"
shodan = "^1.27.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
